+++ psycopg/psycopg/types/array.py
@@ -43,27 +43,43 @@ class BaseListDumper(RecursiveDumper):
             sdclass = context.adapters.get_dumper_by_oid(self.element_oid, self.format)
             self.sub_dumper = sdclass(NoneType, context)
 
-    def _find_list_element(self, L: List[Any]) -> Any:
+    def _find_list_element(self, L: List[Any], format: PyFormat) -> Any:
         """
         Find the first non-null element of an eventually nested list
         """
-        it = self._flatiter(L, set())
-        try:
-            item = next(it)
-        except StopIteration:
+        items = list(self._flatiter(L, set()))
+        types = {type(item): item for item in items}
+        if not types:
             return None
+
+        if len(types) == 1:
+            t, v = types.popitem()
+        else:
+            # More than one type in the list. It might be still good, as long
+            # as they dump with the same oid (e.g. IPv4Network, IPv6Network).
+            dumpers = [self._tx.get_dumper(item, format) for item in types.values()]
+            oids = set(d.oid for d in dumpers)
+            if len(oids) == 1:
+                t, v = types.popitem()
+            else:
+                raise e.DataError(
+                    "cannot dump lists of mixed types;"
+                    f" got: {', '.join(sorted(t.__name__ for t in types))}"
+                )
 
         # Checking for precise type. If the type is a subclass (e.g. Int4)
         # we assume the user knows what type they are passing.
-        if type(item) is not int:
-            return item
+        if t is not int:
+            return v
 
         # If we got an int, let's see what is the biggest one in order to
         # choose the smallest OID and allow Postgres to do the right cast.
-        it = self._flatiter(L, set())
-        imax = max((i if i >= 0 else -i - 1 for i in it), default=0)
-        imax = max(item if item >= 0 else -item - 1, imax)
-        return imax
+        imax: int = max(items)
+        imin: int = min(items)
+        if imin >= 0:
+            return imax
+        else:
+            return max(imax, -imin - 1)
 
     def _flatiter(self, L: List[Any], seen: Set[int]) -> Any:
         if id(L) in seen:
@@ -101,7 +117,7 @@ class ListDumper(BaseListDumper):
         if self.oid:
             return self.cls
 
-        item = self._find_list_element(obj)
+        item = self._find_list_element(obj, format)
         if item is None:
             return self.cls
 
@@ -113,12 +129,12 @@ class ListDumper(BaseListDumper):
         if self.oid:
             return self
 
-        item = self._find_list_element(obj)
+        item = self._find_list_element(obj, format)
         if item is None:
             # Empty lists can only be dumped as text if the type is unknown.
             return self
 
-        sd = self._tx.get_dumper(item, format.from_pq(self.format))
+        sd = self._tx.get_dumper(item, PyFormat.from_pq(self.format))
         dumper = type(self)(self.cls, self._tx)
         dumper.sub_dumper = sd
 
@@ -203,7 +219,7 @@ class ListBinaryDumper(BaseListDumper):
         if self.oid:
             return self.cls
 
-        item = self._find_list_element(obj)
+        item = self._find_list_element(obj, format)
         if item is None:
             return (self.cls,)
 
@@ -215,7 +231,7 @@ class ListBinaryDumper(BaseListDumper):
         if self.oid:
             return self
 
-        item = self._find_list_element(obj)
+        item = self._find_list_element(obj, format)
         if item is None:
             return ListDumper(self.cls, self._tx)