+++ docs/news.rst
@@ -7,6 +7,15 @@
 ``psycopg`` release notes
 =========================
 
+Future releases
+---------------
+
+Psycopg 3.0.14 (unreleased)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+- Raise DataError dumping arrays of mixed types (:ticket:`#301`).
+
+
 Current release
 ---------------